requirements.txt
python-dotenv==0.10.3
Flask==1.1.1
mysql-connector==2.2.9
bcrypt==3.1.7